Support reading dropped columns
The content of dropped columns can be retrieved as long as the table has
not been rewritten (e.g. via `VACUUM FULL` or `CLUSTER`). Use
`dropped_N` to access the Nth column, counting from 1. PostgreSQL
deletes the type information of the original column, so only a few
sanity checks can be done if the correct type was specified in the table
alias; checked are type length, type alignment, type modifier, and
pass-by-value.